requirements.txt
Flask==1.0.2
opencv-python==3.4.0.14
opencv-contrib-python==3.4.0.14
Pillow==5.1.0
Flask-SSlify==0.1.5
pyOpenSSL==18.0.0